send rating information along with messages
This is to help initiating a new conversation threads in this scenario:

A request. B sends A message. A opens chat icon to populate chat list view adapter to create a new thread for this conversation, yet the rating string in this case was "", which causes adapter to signal error.