Use a unique linked list as flipper stack + remove duplicate toolbar title
This is too avoid the annoying situation where the flipperStack gets too large and the user needs to press back many many times inorder to exit app. Using a unique linked list makes sure only maximum 3 entries are in the stack, and the user can exit the app within 3 presses.